requirements.txt.txt
tensorflow
keras
numpy
pillow
cv2
flask
dlib
pylab
flask_cors
